% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22064R.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions (\code{gibbsR} and \code{vaccR}) and Cpp functions (\code{gibbsC} and \code{vaccC}).
}
\examples{
\dontrun{
data(data)
tm1 <- microbenchmark::microbenchmark(shortestR = shortestR(M), shortestC = shortestC(M))
summary(tm1)[,c(1,3,5,6)]

tm2 <- microbenchmark::microbenchmark(gibbR = gibbsR(N=2000, mu1=0, mu2=0, sigma1=1, sigma2=1, pho=0.9), gibbC = gibbsC(100, 0, 0, 1, 1, 0.9))
summary(tm2)[,c(1,3,5,6)]
}
}
