% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22064R.R
\name{data}
\alias{data}
\title{A illustration dataset}
\description{
A dataset used to illustrate the performance of \code{shortestR} and \code{shortestC}.
}
\examples{
\dontrun{
data(data)
tm <- microbenchmark::microbenchmark(
  shortestR = shortestR(M),
  shortestC = shortestC(M)
)
print(summary(tm)[,c(1,3,5,6)])
}
}
