% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shortestC}
\alias{shortestC}
\title{Calculate the shortest path problem between cities using Rcpp.}
\usage{
shortestC(M)
}
\arguments{
\item{M}{the number of the city connected. It's a two-column matrix. The first number of each row is less than the second one.}
}
\value{
a matrix of size \code{max(M)} whose elements are the minimum distances between cities.
}
\description{
The prediction model is described in https://www.math.pku.edu.cn/teachers/lidf/docs/Rbook/html/_Rbook/examples.html#examples-scicomp-citylink.
}
\examples{
\dontrun{
data(data)
distanceC <- shortestC(M)
}
}
