\name{StatComp22064-package}
\alias{StatComp22064-package}
\docType{package}
\title{
  This package provides several functions for benchmarking R and Rcpp.
}
\description{
  R and Rcpp versions of gibbs sampler for a bivariate normal chain, and shortest function for shortest path problem between cities are provided in this package (\code{gibbsR} and \code{shortestR} versus \code{gibbsC} and \code{shortestC}). 
  Besides, all the homework answers for 'Statistical Computing' course of student 22064 are included.
}
\author{
Shiqi Liu

Maintainer: Shiqi Liu <liushiqi@mail.ustc.edu.cn>
}
\keyword{ package }
